+++ packages/winnow/lib/filter-and-transform/transforms/to-esri-geometry.js
@@ -2,9 +2,7 @@ const _ = require('lodash')
 const { geojsonToArcGIS } = require('@terraformer/arcgis')
 
 module.exports = function convert (geometry = {}) {
-  const { type } = geometry
-
-  if (!type) return null
+  if (!geometry || !geometry.type) return null
 
   const result = geojsonToArcGIS(_.clone(geometry))