+++ packages/winnow/test/unit/sql-query-builder/select/geometry-select-fragment.spec.js
@@ -13,8 +13,14 @@ test('createGeometryFragment: empty options', t => {
   t.end()
 })
 
-test('createGeometryFragment: outputCrs option', t => {
-  const geometryFragment = createGeometrySelectFragment({ outputCrs: { wkid: 3857 } })
+test('createGeometryFragment: same input/output CRS option', t => {
+  const geometryFragment = createGeometrySelectFragment({ inputCrs: { wkid: 3857 }, outputCrs: { wkid: 3857 } })
+  t.equal(geometryFragment, 'geometry')
+  t.end()
+})
+
+test('createGeometryFragment: different input/output CRS option', t => {
+  const geometryFragment = createGeometrySelectFragment({ inputCrs: { wkid: 4326 }, outputCrs: { wkid: 3857 } })
   t.equal(geometryFragment, 'project(geometry,?,?) as geometry')
   t.end()
 })
@@ -31,20 +37,20 @@ test('createGeometryFragment: toEsri option', t => {
   t.end()
 })
 
-test('createGeometryFragment: outputCrs and geometryPrecision options', t => {
-  const geometryFragment = createGeometrySelectFragment({ outputCrs: { wkid: 3857 }, geometryPrecision: 1 })
+test('createGeometryFragment: different input/output CRS, and geometryPrecision options', t => {
+  const geometryFragment = createGeometrySelectFragment({ inputCrs: { wkid: 4326 }, outputCrs: { wkid: 3857 }, geometryPrecision: 1 })
   t.equal(geometryFragment, 'reducePrecision(project(geometry,?,?),?) as geometry')
   t.end()
 })
 
-test('createGeometryFragment: outputCrs, geometryPrecision, and toEsri options', t => {
-  const geometryFragment = createGeometrySelectFragment({ outputCrs: { wkid: 3857 }, geometryPrecision: 1, toEsri: true })
+test('createGeometryFragment: different input/output CRS, geometryPrecision, and toEsri options', t => {
+  const geometryFragment = createGeometrySelectFragment({ inputCrs: { wkid: 4326 }, outputCrs: { wkid: 3857 }, geometryPrecision: 1, toEsri: true })
   t.equal(geometryFragment, 'esriGeometry(reducePrecision(project(geometry,?,?),?)) as geometry')
   t.end()
 })
 
-test('createGeometryFragment: outputCrs, and toEsri options', t => {
-  const geometryFragment = createGeometrySelectFragment({ outputCrs: { wkid: 3857 }, toEsri: true })
+test('createGeometryFragment: different input/output CRS, and toEsri options', t => {
+  const geometryFragment = createGeometrySelectFragment({ inputCrs: { wkid: 4326 }, outputCrs: { wkid: 3857 }, toEsri: true })
   t.equal(geometryFragment, 'esriGeometry(project(geometry,?,?)) as geometry')
   t.end()
 })